combined-requirements.txt
# Recommender
beautifulsoup4==4.12.2
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.0
gensim==4.3.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
MarkupSafe==2.1.3
nltk==3.8.1
nodeenv==1.8.0
numpy==1.26.1
pandas==2.1.2
python-dateutil==2.8.2
pytz==2023.3.post1
regex==2023.10.3
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.3
six==1.16.0
smart-open==6.4.0
soupsieve==2.5
threadpoolctl==3.2.0
tqdm==4.66.1
tzdata==2023.3
Unidecode==1.3.7
urllib3==2.0.7
Werkzeug==3.0.1


# Detection
# Base ----------------------------------------
matplotlib>=3.2.2
opencv-python>=4.1.1
Pillow>=7.1.2
PyYAML>=5.3.1
torch>=1.7.0,!=1.12.0
torchvision>=0.8.1,!=0.13.0
protobuf<4.21.3

# Logging -------------------------------------
tensorboard>=2.4.1
# wandb

# Plotting ------------------------------------
seaborn>=0.11.0

# Export --------------------------------------
# coremltools>=4.1  # CoreML export
# onnx>=1.9.0  # ONNX export
# onnx-simplifier>=0.3.6  # ONNX simplifier
# scikit-learn==0.19.2  # CoreML quantization
# tensorflow>=2.4.1  # TFLite export
# tensorflowjs>=3.9.0  # TF.js export
# openvino-dev  # OpenVINO export

# Extras --------------------------------------
ipython  # interactive notebook
psutil  # system utilization
thop  # FLOPs computation
# albumentations>=1.0.3
# pycocotools>=2.0  # COCO mAP
# roboflow
